add jquery and sweetalert2 to code

+++ resources/views/users/index.blade.php
@@ -36,16 +36,25 @@
 @section('javascript')
     $(function (){
         $('.delete').click(function (){
-            $.ajax({
-                method: "DELETE",
-                url: "kuncrog.test/users/" + $(this).data("id"),
-                data: { id: $(this).data("id") }
-            })
-            .done(function( response ) {
-                window.location.reload();
-            })
-            .fail(function (response){
-                alert("error");
+            Swal.fire({
+            title: "Czy na pewno chcesz usunąć rekord?",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonText: 'Tak, usuń',
+            dangerMode: true,
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    $.ajax({
+                        method: "DELETE",
+                        url: "kuncrog.test/users/" + $(this).data("id")
+                    })
+                    .done(function( response ) {
+                        window.location.reload();
+                    })
+                    .fail(function (response){
+                        Swal.fire("Oops!", "Something went wrong!", "error");
+                    });
+                }
             });
         });
     });
